Move all bookings related selectors in "selectors/bookings" folder (including bookings related slots selectors)
